enemy shooting done

+++ Assets/Scripts/HealthBar.cs
@@ -9,9 +9,8 @@ public class HealthBar : MonoBehaviour
     public Gradient gradient;
     public Image fill;
 
-    public void SetHealth(int health) {
+    public void SetHealth(float health) {
         slider.value = health;
-
         fill.color = gradient.Evaluate(slider.normalizedValue);
     }
